requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
pydantic
pydantic-settings
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
aiosqlite
pytest
pytest-asyncio
httpx
python-multipart
